+++ XOSL/IO/Disk.cpp
@@ -100,7 +100,7 @@ int CDisk::Transfer(int Action, unsigned long long Sector, void *Buffer, int Cou
 
 		return DiskAccess.LBATransfer(Action,Drive,LBAPacket);
 	}
-	Sector2CHS(Sector,SectCyl,DrvHead);
+	Sector2CHS((unsigned long)Sector,SectCyl,DrvHead);
 	return DiskAccess.Transfer(Action,SectCyl,DrvHead,Buffer,Count);
 }
 
@@ -111,7 +111,7 @@ void CDisk::Sector2CHS(unsigned long RSector, unsigned short &SectCyl, unsigned 
 	int Sector;
 	int Head;
 
-	RSector += StartSector;
+	RSector += (unsigned long)StartSector;
 
 	Sector = (int)(RSector % DrvSectorCount + 1);
 	RSector /= DrvSectorCount;
